+++ docs/versionhistory.rst
@@ -8,6 +8,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Fixed race condition in ``Lock`` and ``Semaphore`` classes when a task waiting on ``acquire()``
   is cancelled while another task is waiting to acquire the same primitive
   (`#387 <https://github.com/agronholm/anyio/issues/387>`_)
+- Fixed context variables not propagating to worker threads in ``to_thread.run_sync()``
+  (partially fixes `#363 <https://github.com/agronholm/anyio/issues/363>`_)
 
 **3.3.4**